+++ transformer/plugins/__init__.py
@@ -1,5 +1,0 @@
-from typing import Sequence, Callable
-
-from transformer.task import Task
-
-Plugin = Callable[[Sequence[Task]], Sequence[Task]]