+++ transformer/transform.py
@@ -7,17 +7,19 @@ from typing import Sequence, Union
 
 import transformer.python as py
 from transformer.locust import locustfile
-from transformer.plugins import Plugin
+from transformer.plugins.contracts import OnTaskSequence
 from transformer.scenario import Scenario
 
 
 def transform(
-    scenarios_path: Union[str, Path], plugins: Sequence[Plugin] = ()
+    scenarios_path: Union[str, Path], plugins: Sequence[OnTaskSequence] = ()
 ) -> py.Program:
     return locustfile([Scenario.from_path(Path(scenarios_path), plugins)])
 
 
-def main(scenarios_path: Union[str, Path], plugins: Sequence[Plugin] = ()) -> str:
+def main(
+    scenarios_path: Union[str, Path], plugins: Sequence[OnTaskSequence] = ()
+) -> str:
     """
     Converts a WeightedFilePaths into a Locustfile.
     """