+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Plugin contracts aiming to eventually deprecate OnTaskSequence:
+  OnTask, OnScenario, OnPythonProgram.
+
 ### Changed
 
 - Open-sourcing of this project in https://github.com/zalando-incubator.
+- `transformer.plugins.Plugin` is renamed
+  `transformer.plugins.contracts.OnTaskSequence`.
 
 [Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.0...HEAD